+++ src/game/imagelayer.cpp
@@ -18,7 +18,7 @@ std::shared_ptr<ImageLayer> ImageLayer::deserialize(const std::shared_ptr<TmxEle
   image->_z_index = image_layer->_z;
   image->_texture = TexturePool::getInstance().get((level_path / image_layer->_image->_source).string());
   image->_sprite.setPosition(image_layer->_offset_x_px, image_layer->_offset_y_px);
-  image->_sprite.setColor(sf::Color(255, 255, 255, static_cast<uint32_t>(image_layer->_opacity * 255.0f)));
+  image->_sprite.setColor(sf::Color(255, 255, 255, static_cast<uint8_t>(image_layer->_opacity * 255.0f)));
   image->_sprite.setTexture(*image->_texture);
 
   sf::BlendMode blend_mode = sf::BlendAdd;
